Add toArrays, fromSortedArrays, and size
The implementation of size is straightforward and has the abysmal
performance that would be expected for a B-Tree. The implementation
of toArrays is similarly straightforward, but it is a high-performance
transformation useful for dumping B-Trees out as arrays so that filters
or traversals can be applied. Finally, fromSortedArrays is the reverse
of toArrays. Its implementation is unfortunately a little complicated,
but it has been thoroughly tested. Its results always have full branches
and leaves, a desirable property, and it operates in O(n) time.